requirements_demo.txt
absl-py==0.10.0
altair==4.1.0
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
audioread==2.1.8
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.15.7
botocore==1.18.7
bert-extractive-summarizer==0.7.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
entrypoints==0.3
enum-compat==0.0.3
ffmpeg==1.4
future==0.18.2
gast==0.2.2
glob2==0.7
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
inflect==4.1.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.31.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
multiprocess==0.70.10
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numba==0.48.0
numpy==1.19.2
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pooch==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==1.0.1
pybase64==1.1.4
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.7.1
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
pyzmq==19.0.2
regex==2021.3.17
requests==2.24.0
resampy==0.2.2
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
streamlit==0.67.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.11.1
torch==1.6.0
torchfile==0.1.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.50.0
traitlets==5.0.4
tzlocal==2.1
umap-learn==0.4.6
Unidecode==1.1.1
urllib3==1.25.10
validators==0.18.1
visdom==0.1.8.9
watchdog==0.10.3
wavio==0.0.4
wcwidth==0.2.5
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.2.0
